+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@fontsource/roboto": "^5.0.8",
+    "clsx": "^2.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
