+++ src/assets/icons/LogOutIcon.module.scss
@@ -1,0 +1,5 @@
+.svg {
+  flex-shrink: 0;
+  width: 16px;
+  height: 16px;
+}
