+++ src/components/ui/Button/Button.module.scss
@@ -4,6 +4,7 @@
   cursor: pointer;
 
   display: flex;
+  gap: 10px;
   align-items: center;
   justify-content: center;
 
@@ -21,7 +22,7 @@
   border-radius: 4px;
 
   svg {
-    fill: red;
+    fill: currentcolor;
   }
 
   &:focus-visible {
@@ -85,8 +86,12 @@
       outline: none;
     }
   }
+}
 
-  .fullWidth {
-    width: 100%;
-  }
+.fullWidth {
+  width: 100%;
+}
+
+a.button {
+  display: inline-flex;
 }
